# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TBH VIP Telegram Bot + Admin Panel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Panel {
  id: number;
  name: string;
  firebaseUrl: string;
  createdAt: string;
}

export interface PanelInput {
  name: string;
  firebaseUrl: string;
  secretKey: string;
}

export interface Device {
  id: string;
  name: string;
  status: boolean;
  phoneNumber: string;
  battery: string;
  model?: string;
  /** @nullable */
  lastSeen?: string | null;
  simCount?: number;
  totalSms?: number;
  panelId?: number;
  panelName?: string;
}

export interface PanelStat {
  panelId: number;
  panelName: string;
  online: number;
  offline: number;
  total: number;
}

export interface DashboardStats {
  totalPanels: number;
  totalDevices: number;
  onlineDevices: number;
  offlineDevices: number;
  totalUsers: number;
  activeUsers: number;
  totalGiftCards: number;
  usedGiftCards: number;
  panelBreakdown?: PanelStat[];
}

export interface BotUser {
  id: number;
  telegramId: string;
  /** @nullable */
  username?: string | null;
  firstName: string;
  referralCode: string;
  /** @nullable */
  referredBy?: string | null;
  referralCount: number;
  smsCredits: number;
  /** @nullable */
  getNumberExpiresAt?: string | null;
  sendSmsUnlocked?: boolean;
  /** @nullable */
  webPanelExpiresAt?: string | null;
  createdAt: string;
}

export interface UserUpdate {
  smsCredits?: number;
  /** @nullable */
  getNumberExpiresAt?: string | null;
  sendSmsUnlocked?: boolean;
  /** @nullable */
  webPanelExpiresAt?: string | null;
}

export type GiftCardType = typeof GiftCardType[keyof typeof GiftCardType];


export const GiftCardType = {
  hours: 'hours',
  credits: 'credits',
} as const;

export interface GiftCard {
  id: number;
  code: string;
  type: GiftCardType;
  value: number;
  /** @nullable */
  usedBy?: string | null;
  /** @nullable */
  usedAt?: string | null;
  createdAt: string;
}

export type GiftCardInputType = typeof GiftCardInputType[keyof typeof GiftCardInputType];


export const GiftCardInputType = {
  hours: 'hours',
  credits: 'credits',
} as const;

export interface GiftCardInput {
  code: string;
  type: GiftCardInputType;
  value: number;
}

export interface TelegramUpdate {
  update_id?: number;
}

export type ListUsersParams = {
search?: string;
};