# lib/api-zod/src/generated/types/giftCard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TBH VIP Telegram Bot + Admin Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { GiftCardType } from './giftCardType';

export interface GiftCard {
  id: number;
  code: string;
  type: GiftCardType;
  value: number;
  /** @nullable */
  usedBy?: string | null;
  /** @nullable */
  usedAt?: string | null;
  createdAt: string;
}