# lib/api-zod/src/generated/types/giftCardInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TBH VIP Telegram Bot + Admin Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { GiftCardInputType } from './giftCardInputType';

export interface GiftCardInput {
  code: string;
  type: GiftCardInputType;
  value: number;
}