# lib/api-zod/src/generated/types/telegramUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TBH VIP Telegram Bot + Admin Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface TelegramUpdate {
  update_id?: number;
}